requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0